Move BackgroundKeepAlive class to its own file
